Reduced duplicate code
