+++ api/src/main/java/io/kafbat/ui/serdes/builtin/StructSerde.java
@@ -5,12 +5,16 @@ import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
+import io.kafbat.ui.serde.api.SchemaDescription;
+import io.kafbat.ui.serde.api.Serde;
+import io.kafbat.ui.serdes.BuiltInSerde;
 import java.io.IOException;
+import java.util.Optional;
 import lombok.SneakyThrows;
 import org.apache.kafka.common.protocol.types.BoundField;
 import org.apache.kafka.common.protocol.types.Struct;
 
-public abstract class StructSerde {
+public abstract class StructSerde implements BuiltInSerde {
 
   private static final JsonMapper JSON_MAPPER = createMapper();
 
@@ -30,6 +34,31 @@ public abstract class StructSerde {
     var mapper = new JsonMapper();
     mapper.registerModule(module);
     return mapper;
+  }
+
+  @Override
+  public Optional<String> getDescription() {
+    return Optional.empty();
+  }
+
+  @Override
+  public Optional<SchemaDescription> getSchema(String topic, Serde.Target type) {
+    return Optional.empty();
+  }
+
+  @Override
+  public boolean canDeserialize(String topic, Serde.Target type) {
+    return true;
+  }
+
+  @Override
+  public boolean canSerialize(String topic, Serde.Target type) {
+    return false;
+  }
+
+  @Override
+  public Serde.Serializer serializer(String topic, Serde.Target type) {
+    throw new UnsupportedOperationException();
   }
 
   @SneakyThrows
